WIP Edit Flex Image  - Image Chooser implementation, working better on chrome.  the dataURL might be problematic on FF  - EditFlexImageViewModel needs work

+++ SnailbirdMedia/Models/MediaBinaryDto.cs
@@ -3,7 +3,7 @@
     internal class MediaBinaryDto
     {
         public byte[] Bytes { get; set; } = default!;
-        public int Size { get; set; } = default!;
+        public long Size { get; set; } = default!;
         public string Extension { get; set; } = default!;
     }
 }
